adjust Qt GUI data types

+++ src/age_qt_gui/age_ui_qt.hpp
@@ -42,14 +42,11 @@ std::string operator+(const std::string &std_string, const QString &q_string);
 namespace age
 {
 
+//! \todo refactor & remove
 typedef std::atomic_size_t  atomic_uint;
 typedef std::atomic<uint64> atomic_uint64;
 
-
-
 bool is_checked(int checked_state);
-
-
 
 
 
@@ -64,10 +61,8 @@ bool is_checked(int checked_state);
 //!
 //! While in theory we could handle bigger files, it does not make sense for our use cases.
 //! There are no cartridge or RAM files bigger than this.
-//! In fact it could even lower performance and stability to try loading such big files
-//! (think of a RAM file of several gigabytes being loaded into memory).
 //!
-constexpr qint64 max_file_bytes = 32 * 1024 * 1024;
+constexpr int max_file_bytes = 32 * 1024 * 1024;
 
 
 
@@ -113,9 +108,7 @@ qt_downsampler_quality get_qt_downsampler_quality_for_name(const QString &qualit
 //
 //---------------------------------------------------------
 
-constexpr uint qt_video_frame_history_size = 4;
-
-
+constexpr int qt_video_frame_history_size = 4;
 
 // post processing filter
